+++ packages/vue/src/components/timer/use-timer-context.ts
@@ -1,0 +1,6 @@
+import { createContext } from '../../utils'
+import type { UseTimerReturn } from './use-timer'
+
+export interface UseTimerContext extends UseTimerReturn {}
+
+export const [TimerProvider, useTimerContext] = createContext<UseTimerContext>('TimerContext')